Add get endpoint test and properly configure test
