+++ app.js
@@ -1,0 +1,22 @@
+const environment = process.env.NODE_ENV || 'development';
+const configuration = require('./knexfile')[environment];
+const database = require('knex')(configuration);
+const express = require('express');
+const app = express();
+app.use(express.json());
+const bodyParser = require('body-parser')
+app.use(bodyParser.json());
+const cors = require('cors')
+app.use(cors())
+
+app.get('/api/v1/projects', (request, response) => {
+  database('projects').select()
+    .then(projects => {
+      response.status(200).json(projects)
+    })
+    .catch(error => {
+      response.status(500).json({ error })
+    })
+});
+
+module.exports = app
