+++ package.json
@@ -11,8 +11,8 @@
   "license": "ISC",
   "dependencies": {
     "@babel/core": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
     "babel-jest": "^24.8.0",
-    "babel-preset-env": "^1.7.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "express": "^4.16.4",
