+++ server.js
@@ -1,22 +1,9 @@
-const environment = process.env.NODE_ENV || 'development';
-const configuration = require('./knexfile')[environment];
-const database = require('knex')(configuration);
-const express = require('express');
-const bodyParser = require('body-parser')
-const app = express();
-app.use(bodyParser.json());
-app.set('port', process.env.PORT || 3001);
+const app = require('./app')
 
-app.get('/api/v1/projects', (request, response) => {
-  database('projects').select()
-    .then(projects => {
-      response.status(200).json(projects)
-    })
-    .catch(error => {
-      response.status(500).json({ error })
-    })
-});
+app.set('port', process.env.PORT || 3001);
 
 app.listen(app.get('port'), () => {
   console.log(`App is running on localhost: ${app.get('port')}.`)
 });
+
+module.exports = app
